requirements.txt
Django==1.8.6
djangorestframework==3.3.1
markdown==2.6.5
django-filter=0.=11.0